+++ plugins/noaa-dsb/src/deframer.h
@@ -32,5 +32,5 @@ public:
     // Return state
     int getState();
     // Perform deframing
-    std::vector<std::array<uint8_t, FRAME_SIZE>> work(uint8_t *input, size_t size);
+    std::vector<std::array<uint8_t, FRAME_SIZE>> work(uint8_t *input, int size);
 };
